Create Marketing Analysis With Clustering.ipynb
Customer segmentation is always the fascinating problem for every business. For the retailer, they will want to understand the shopping behavior of their customer so that they could have the better strategies for each group of customer. Therefore, more companies have been interested in analyzing and clustering their customers in order to market the new products to exact the customer who needs. In this project, I will introduce the most common technique in segmentation is cluster analysis. This help us explore the similarity between each customer and then group them as the same cluster who share the same behavior